Make requested changes and clean up some stuff
