+++ core/src/channel_group/mod.rs
@@ -27,26 +27,33 @@ pub struct ChannelGroup {
 /// Options for initializing a new ChannelGroup.
 pub struct ChannelGroupConfig {
     /// Channel initialization options (same for all channels).
+    /// See the `ChannelInitOptions` documentation for more information.
     pub channel_init_options: ChannelInitOptions,
 
-    /// Amount of VoiceChannel objects to be created.
+    /// Amount of VoiceChannel objects to be created
     /// (Number of MIDI channels)
+    /// The MIDI 1 spec uses 16 channels.
     pub channel_count: u32,
 
     /// A vector which specifies which of the created channels (indexes) will be used for drums.
+    ///
     /// For example in a conventional 16 MIDI channel setup where channel 10 is used for
-    /// drums, the vector would be set as \[9\] (counting from 0).
+    /// drums, the vector would be set as vec!\[9\] (counting from 0).
     pub drums_channels: Vec<u32>,
 
     /// Parameters of the output audio.
+    /// See the `AudioStreamParams` documentation for more information.
     pub audio_params: AudioStreamParams,
 
-    /// Whether or not to use a threadpool to render voices.
+    /// Whether or not to use a threadpool to render individual voices.
+    /// Regardless, each MIDI channel uses its own thread. This setting
+    /// adds more fine-grained threading per key rather than per channel.
     pub use_threadpool: bool,
 }
 
 impl ChannelGroup {
     /// Creates a new ChannelGroup with the given configuration.
+    /// See the `ChannelGroupConfig` documentation for the available options.
     pub fn new(config: ChannelGroupConfig) -> Self {
         let mut channels = Vec::new();
         let mut channel_events_cache = Vec::new();
@@ -82,7 +89,7 @@ impl ChannelGroup {
     }
 
     /// Sends a SynthEvent to the ChannelGroup.
-    /// Please see the SynthEvent documentation for more information.
+    /// See the `SynthEvent` documentation for more information.
     pub fn send_event(&mut self, event: SynthEvent) {
         match event {
             SynthEvent::Channel(channel, event) => {
@@ -153,7 +160,7 @@ impl ChannelGroup {
         });
     }
 
-    /// Returns the active voice count of all the MIDI channels.
+    /// Returns the active voice count of the synthesizer.
     pub fn voice_count(&self) -> u64 {
         self.channels
             .iter()
